qemux86/qemuarm: Drop kernel version overrides
Drop the version overrides for the kernel for the x86 and arm machines
so we can go back to following the distro versions. The reasons for
these versions is mostly historical at this point as the issues were
resolved.

(From OE-Core rev: 12d174d0b095288533d6d38f622b707f6111e5d0)

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>